Rename winBindings.CommCtrl to winBindings.CommCtrl
